# Add arbitrary telemetry client package, properties