Add a package for arbitrary events that can be imported separately, avoiding importing the config system and counter APIs, and avoiding exposing the arbitrary event APIs to the Microsoft build of Go.

Add support for `Properties` to the events. This is pretty simple: just an extra map pointer that's always nil for the Microsoft build of Go cases.

Includes properties map sanitization, copied from the original module. Made a simple test to check that it's working.

Doesn't include features like CommonProperties attached to the Client--this can be done just as well (and perhaps better) by the caller. No metrics, only events.

No effect on the benchmark:

```
goos: windows
goarch: amd64
pkg: github.com/microsoft/go-infra/telemetry/internal/appinsights
cpu: Intel(R) Core(TM) i7-9700 CPU @ 3.00GHz
                         │ before.txt  │           after.txt           │
                         │   sec/op    │   sec/op     vs base          │
ClientBurstPerformance-8   695.6n ± 2%   693.4n ± 2%  ~ (p=0.853 n=10)
```